fix: invalid product images
As stated by Homey:
> — The app image is identical to the driver image, this is not allowed.
Please add a unique app image that is visually appealing and represent
the brand.
For apps that support a specific brand often lifestyle image of the
brand are used. Consider an image like this:
https://wallbox.com/media_usa/cms/home/business_slide_2.jpg

For more information and examples have a look at our App Store
Guidelines section 1.4.2. App Images.
https://apps.developer.homey.app/app-store/guidelines#1.4.-images